fix: MapFragment outside LocationFragment
