+++ app/src/main/java/me/andreandyp/androidtechnicalchallenge/LocationFragment.kt
@@ -3,7 +3,6 @@ package me.andreandyp.androidtechnicalchallenge
 import android.app.AlertDialog
 import android.content.Context
 import android.content.DialogInterface
-import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -13,28 +12,18 @@ import android.widget.ArrayAdapter
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.Polygon
-import com.google.android.gms.maps.model.PolygonOptions
 import com.google.android.material.textfield.TextInputEditText
 import me.andreandyp.androidtechnicalchallenge.databinding.LocationFragmentBinding
 import me.andreandyp.androidtechnicalchallenge.network.NetworkResponse
 import me.andreandyp.androidtechnicalchallenge.repository.models.ZipCodeSettlements
 
 class LocationFragment : Fragment() {
-
     private lateinit var viewModel: LocationViewModel
-    private lateinit var mapPolygon: SupportMapFragment
     private lateinit var binding: LocationFragmentBinding
     private lateinit var loadingPolygonsDialog: AlertDialog
     private lateinit var loadingSettlementsDialog: AlertDialog
     private lateinit var errorDialog: AlertDialog
     private lateinit var errorNetworkDialog: AlertDialog
-    private var map: GoogleMap? = null
-    private var currentPolygon: Polygon? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -42,12 +31,6 @@ class LocationFragment : Fragment() {
     ): View {
         binding = DataBindingUtil.inflate(inflater, R.layout.location_fragment, container, false)
         binding.lifecycleOwner = this
-        mapPolygon = childFragmentManager.findFragmentById(R.id.map_polygon) as SupportMapFragment
-
-        mapPolygon.getMapAsync { googleMap: GoogleMap ->
-            map = googleMap
-            viewModel.onReadyMap()
-        }
         loadingPolygonsDialog = createLoadingDialog(R.layout.loading_polygon_dialog)
         loadingSettlementsDialog = createLoadingDialog(R.layout.loading_settlements_dialog)
         return binding.root
@@ -55,14 +38,8 @@ class LocationFragment : Fragment() {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(LocationViewModel::class.java)
+        viewModel = ViewModelProvider(requireActivity()).get(LocationViewModel::class.java)
         binding.vm = viewModel
-
-        viewModel.geoPoints.observe(viewLifecycleOwner) { updatePolygon(it) }
-
-        viewModel.centerPoint.observe(viewLifecycleOwner) { point: LatLng ->
-            map?.moveCamera(CameraUpdateFactory.newLatLngZoom(point, 14f))
-        }
 
         viewModel.zipCode.observe(viewLifecycleOwner) { getZipCodeInformation(it) }
 
@@ -88,17 +65,6 @@ class LocationFragment : Fragment() {
                 }
             }
         }
-    }
-
-    private fun updatePolygon(points: List<LatLng>) {
-        currentPolygon?.remove()
-        val polygonOptions = PolygonOptions().run {
-            addAll(points)
-            strokeColor(Color.BLACK)
-            fillColor(R.color.black_light)
-            geodesic(true)
-        }
-        currentPolygon = map?.addPolygon(polygonOptions)
     }
 
     private fun updateSettlements(settlements: ZipCodeSettlements) {
@@ -141,11 +107,11 @@ class LocationFragment : Fragment() {
 
     private fun getZipCodeInformation(zipCode: String) {
         if (zipCode.length == 5) {
+            viewModel.getPolygonCoordinates(zipCode)
+            viewModel.getPolygonsDetails(zipCode)
             val imm: InputMethodManager =
                 requireActivity().getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
             imm.hideSoftInputFromWindow(requireActivity().currentFocus?.windowToken, 0)
-            viewModel.getPolygonCoordinates(zipCode)
-            viewModel.getPolygonsDetails(zipCode)
         }
     }
 }
