+++ app/src/main/res/layout/location_fragment.xml
@@ -139,19 +139,5 @@
                 android:hint="@string/settlement"
                 android:inputType="textPersonName" />
         </com.google.android.material.textfield.TextInputLayout>
-
-        <fragment
-            android:id="@+id/map_polygon"
-            android:name="com.google.android.gms.maps.SupportMapFragment"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_marginStart="@dimen/horizontal_margin"
-            android:layout_marginTop="@dimen/vertical_padding"
-            android:layout_marginEnd="@dimen/horizontal_margin"
-            android:layout_marginBottom="@dimen/vertical_padding"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/et_municipality" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
